Switch back to usize-Extensions

+++ src/extensions.rs
@@ -2,7 +2,7 @@ use serde_derive::{Deserialize, Serialize};
 
 bitflags::bitflags! {
     #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
-    pub struct Extensions: u64 {
+    pub struct Extensions: usize {
         const UNWRAP_NEWTYPES = 0x1;
         const IMPLICIT_SOME = 0x2;
         const UNWRAP_VARIANT_NEWTYPES = 0x4;
